refactor, test: metric emission for imbalance inside lb package, add test
