+++ service/sharddistributor/leader/process/processor.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"fmt"
 	"maps"
-	"math"
 	"math/rand"
 	"slices"
 	"sort"
@@ -467,8 +466,7 @@ func (p *namespaceProcessor) rebalanceShardsImpl(ctx context.Context, metricsLoo
 	isRebalancedByShardLoad := len(loadBalanceMoves) > 0
 
 	p.emitExecutorMetric(namespaceState, metricsLoopScope)
-	emitSmoothedLoadMetrics := p.sdConfig.GetLoadBalancingMode(p.namespaceCfg.Name) == types.LoadBalancingModeGREEDY
-	p.emitAssignmentImbalanceMetrics(metricsLoopScope, currentAssignments, namespaceState, emitSmoothedLoadMetrics)
+	loadbalancer.EmitAssignmentImbalanceMetrics(p.sdConfig, p.namespaceCfg.Name, metricsLoopScope, currentAssignments, namespaceState)
 
 	distributionChanged := len(deletedShards) > 0 || len(staleExecutors) > 0 || assignedToEmptyExecutors || updatedAssignments || isRebalancedByShardLoad
 	if !distributionChanged {
@@ -817,119 +815,4 @@ func makeShards(num int64) []string {
 		shards[i] = strconv.FormatInt(i, 10)
 	}
 	return shards
-}
-
-func (p *namespaceProcessor) emitAssignmentImbalanceMetrics(
-	metricsLoopScope metrics.Scope,
-	assignments map[string][]string,
-	namespaceState *store.NamespaceState,
-	emitSmoothedLoadMetrics bool,
-) {
-	if metricsLoopScope == nil || namespaceState == nil {
-		return
-	}
-
-	reportedLoads := make([]float64, 0, len(assignments))
-	var smoothedLoads []float64
-	if emitSmoothedLoadMetrics {
-		smoothedLoads = make([]float64, 0, len(assignments))
-	}
-
-	totalAssigned := 0
-	smoothedMissing := 0
-
-	for executorID, shards := range assignments {
-		reportedLoad := 0.0
-		smoothedLoad := 0.0
-
-		heartbeat, heartbeatOK := namespaceState.Executors[executorID]
-		for _, shardID := range shards {
-			totalAssigned++
-
-			if !heartbeatOK || heartbeat.ReportedShards == nil {
-				continue
-			} else if shardReport, ok := heartbeat.ReportedShards[shardID]; ok && shardReport != nil {
-				reportedLoad += shardReport.ShardLoad
-			}
-
-			if !emitSmoothedLoadMetrics {
-				continue
-			}
-			if namespaceState.ShardStats == nil {
-				smoothedMissing++
-				continue
-			}
-			stats, ok := namespaceState.ShardStats[shardID]
-			if !ok || stats.LastUpdateTime.IsZero() {
-				smoothedMissing++
-				continue
-			}
-			smoothedLoad += stats.SmoothedLoad
-		}
-
-		reportedLoads = append(reportedLoads, reportedLoad)
-		if emitSmoothedLoadMetrics {
-			smoothedLoads = append(smoothedLoads, smoothedLoad)
-		}
-	}
-
-	metricsLoopScope.UpdateGauge(metrics.ShardDistributorAssignmentLoadMaxOverMean, maxOverMean(reportedLoads))
-	metricsLoopScope.UpdateGauge(metrics.ShardDistributorAssignmentLoadCV, coefficientOfVariation(reportedLoads))
-	if !emitSmoothedLoadMetrics {
-		return
-	}
-
-	metricsLoopScope.UpdateGauge(metrics.ShardDistributorAssignmentSmoothedLoadMaxOverMean, maxOverMean(smoothedLoads))
-	metricsLoopScope.UpdateGauge(metrics.ShardDistributorAssignmentSmoothedLoadCV, coefficientOfVariation(smoothedLoads))
-
-	if totalAssigned == 0 {
-		metricsLoopScope.UpdateGauge(metrics.ShardDistributorAssignmentSmoothedLoadMissingRatio, 0)
-		return
-	}
-
-	metricsLoopScope.UpdateGauge(metrics.ShardDistributorAssignmentSmoothedLoadMissingRatio, float64(smoothedMissing)/float64(totalAssigned))
-}
-
-func maxOverMean(values []float64) float64 {
-	if len(values) == 0 {
-		return 0
-	}
-
-	total := 0.0
-	maxValue := 0.0
-	for _, value := range values {
-		total += value
-		if value > maxValue {
-			maxValue = value
-		}
-	}
-	mean := total / float64(len(values))
-	if mean == 0 {
-		return 0
-	}
-	return maxValue / mean
-}
-
-func coefficientOfVariation(values []float64) float64 {
-	if len(values) == 0 {
-		return 0
-	}
-
-	total := 0.0
-	for _, value := range values {
-		total += value
-	}
-	mean := total / float64(len(values))
-	if mean == 0 {
-		return 0
-	}
-
-	variance := 0.0
-	for _, value := range values {
-		delta := value - mean
-		variance += delta * delta
-	}
-	variance /= float64(len(values))
-
-	return math.Sqrt(variance) / mean
 }
